Moved FormatRelation and FormatWrapper to the proper directory.
